# Improve performance and reduce allocations of RAMJobStore.StoreJobInternal when job exists