Eliminate extra lookup in `RAMJobStore.StoreJobInternal(IJobDetail new…Job, bool replaceExisting)` when **replaceExisting** is **true** and job exists.

Only create **JobWrapper** when job doesn't yet exist.

**Before:**

|                                      Method |       Mean |     Error |    StdDev |     Median | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|-------------------------------------------- |-----------:|----------:|----------:|-----------:|------------:|------------:|------------:|--------------------:|
|   StoreJob_ReplaceExisting_JobExists_Serial |   275.5 ns | 11.284 ns | 31.456 ns |   266.6 ns |      0.0658 |           - |           - |               104 B |
| StoreJob_ReplaceExisting_JobExists_Parallel |   657.8 ns |  3.677 ns |  3.439 ns |   657.7 ns |      0.0758 |      0.0008 |      0.0008 |                 1 B |
|   StoreJob_ReplaceExisting_JobDoesNotExists | 1,516.1 ns | 30.047 ns | 39.069 ns | 1,510.5 ns |      0.5989 |           - |           - |               944 B |

**After:**

|                                      Method |       Mean |     Error |    StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|-------------------------------------------- |-----------:|----------:|----------:|------------:|------------:|------------:|--------------------:|
|   StoreJob_ReplaceExisting_JobExists_Serial |   147.6 ns | 0.3402 ns | 0.2841 ns |      0.0455 |           - |           - |                72 B |
| StoreJob_ReplaceExisting_JobExists_Parallel |   632.4 ns | 8.9159 ns | 7.4452 ns |      0.0563 |      0.0008 |      0.0008 |                 1 B |
|   StoreJob_ReplaceExisting_JobDoesNotExists | 1,503.2 ns | 7.8769 ns | 6.5776 ns |      0.5989 |           - |           - |               944 B |

Benchmark source is available [here](https://gist.github.com/drieseng/7c2b6e8967f01a72d8b96458dff8a5ac).